- Moving files around. - Cleaning up custom files.
