+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "workbench.editor.enablePreview": true,
+    "editor.rename.enablePreview": true
+}
