+++ OS X/.gitignore
@@ -144,3 +144,10 @@ handshakes/pcap/*
 
 # attacks
 attacks_list.js
+
+# pacakge
+package-lock.json
+
+# Wordlists
+wordlists/known-wpa-passwords.txt
+wordlists/known-wpa-passwords.dic
