+++ OS X/pwnagetty/package.json
