+++ OS X/scripts/network-cracked-status.json
