--- OS X/wordlists/known-wpa-passwords.txt
@@ -1,1 +1,0 @@
-password
